Fixed issue with copying data from old to new
When you'd copy (create-new-from-old) a report, it looked like the copy
was being created properly. However, when you return to the original
report you copied, it would load the data members of the copy (name, and
some values being updated only on create).

What I found to actually be happening is that when we created a copy, we
were actually setting its data members to _references_ to the original
report's data members. This is why the original's data members would
have been updated by a function that shouldn't have been firing when
loading that report; it wasn't, it was just just changing the original
from the copy.

I hate JavaScript.

Note, there are a few ways to properly copy an object, we had to use a
deep copy method that used `JSON.parse` and `JSON.stringify`. I tried
Object.assign without any luck.

[Reference](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign)